Refactor: break up measurables pages
