+++ lotti/lib/pages/settings/measurables/measurable_detail_page.dart
@@ -1,214 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:lotti/classes/entity_definitions.dart';
-import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/pages/settings/form_text_field.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/utils/file_utils.dart';
-import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 
 const double iconSize = 24.0;
-
-class MeasurablesPage extends StatefulWidget {
-  const MeasurablesPage({Key? key}) : super(key: key);
-
-  @override
-  State<MeasurablesPage> createState() => _MeasurablesPageState();
-}
-
-class _MeasurablesPageState extends State<MeasurablesPage> {
-  final JournalDb _db = getIt<JournalDb>();
-
-  late final Stream<List<MeasurableDataType>> stream =
-      _db.watchMeasurableDataTypes();
-
-  @override
-  void initState() {
-    super.initState();
-    createDefaults();
-  }
-
-  void createDefaults() async {
-    DateTime now = DateTime.now();
-
-    _db.upsertMeasurableDataType(MeasurableDataType(
-      id: '9e9e7a62-1e56-4059-a568-12234db7399b',
-      createdAt: now,
-      updatedAt: now,
-      name: 'water',
-      displayName: 'Water',
-      description: 'Volume of water consumed, in milliliters',
-      unitName: 'ml',
-      version: 0,
-      vectorClock: null,
-    ));
-
-    _db.upsertMeasurableDataType(MeasurableDataType(
-      id: 'f2518f33-af1d-4dbe-ae9b-6a05def5d8f9',
-      createdAt: now,
-      updatedAt: now,
-      name: 'caffeine',
-      displayName: 'Caffeine',
-      description: 'Amount of caffeine consumed, in milligrams',
-      unitName: 'mg',
-      version: 0,
-      vectorClock: null,
-    ));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return StreamBuilder<List<MeasurableDataType>>(
-      stream: stream,
-      builder: (
-        BuildContext context,
-        AsyncSnapshot<List<MeasurableDataType>> snapshot,
-      ) {
-        List<MeasurableDataType> items = snapshot.data ?? [];
-
-        return Stack(
-          children: [
-            ListView(
-              shrinkWrap: true,
-              padding: const EdgeInsets.all(8.0),
-              children: List.generate(
-                items.length,
-                (int index) {
-                  return MeasurableTypeCard(
-                    item: items.elementAt(index),
-                    index: index,
-                  );
-                },
-              ),
-            ),
-            Align(
-              alignment: Alignment.bottomRight,
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: FloatingActionButton(
-                  child: const Icon(MdiIcons.plus, size: 32),
-                  backgroundColor: AppColors.entryBgColor,
-                  onPressed: () {
-                    Navigator.of(context).push(
-                      MaterialPageRoute(
-                        builder: (BuildContext context) {
-                          DateTime now = DateTime.now();
-                          return DetailRoute(
-                            item: MeasurableDataType(
-                              id: uuid.v1(),
-                              name: '',
-                              displayName: '',
-                              version: 0,
-                              createdAt: now,
-                              updatedAt: now,
-                              unitName: '',
-                              description: '',
-                              vectorClock: null,
-                            ),
-                            index: -1,
-                          );
-                        },
-                      ),
-                    );
-                  },
-                ),
-              ),
-            )
-          ],
-        );
-      },
-    );
-  }
-}
-
-class MeasurableTypeCard extends StatelessWidget {
-  final MeasurableDataType item;
-  final int index;
-
-  const MeasurableTypeCard({
-    Key? key,
-    required this.item,
-    required this.index,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(4.0),
-      child: Card(
-        color: AppColors.headerBgColor,
-        elevation: 8.0,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(8.0),
-        ),
-        child: SingleChildScrollView(
-          child: ListTile(
-            contentPadding:
-                const EdgeInsets.only(left: 24, top: 4, bottom: 12, right: 24),
-            title: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: [
-                Text(
-                  item.displayName,
-                  style: TextStyle(
-                    color: AppColors.entryTextColor,
-                    fontFamily: 'Oswald',
-                    fontSize: 24.0,
-                  ),
-                ),
-                Expanded(child: Container()),
-                Visibility(
-                  visible: fromNullableBool(item.private),
-                  child: Icon(
-                    MdiIcons.security,
-                    color: AppColors.error,
-                    size: iconSize,
-                  ),
-                ),
-                Visibility(
-                  visible: fromNullableBool(item.favorite),
-                  child: Padding(
-                    padding: const EdgeInsets.only(left: 4.0),
-                    child: Icon(
-                      MdiIcons.star,
-                      color: AppColors.starredGold,
-                      size: iconSize,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            subtitle: Text(
-              item.description,
-              style: TextStyle(
-                color: AppColors.entryTextColor,
-                fontFamily: 'Oswald',
-                fontWeight: FontWeight.w200,
-                fontSize: 16.0,
-              ),
-            ),
-            enabled: true,
-            onTap: () {
-              Navigator.of(context).push(
-                MaterialPageRoute(
-                  builder: (BuildContext context) {
-                    return DetailRoute(
-                      item: item,
-                      index: index,
-                    );
-                  },
-                ),
-              );
-            },
-          ),
-        ),
-      ),
-    );
-  }
-}
 
 class DetailRoute extends StatefulWidget {
   const DetailRoute({
